api/requirements.txt
flask==3.1.0
python-dotenv==1.0.1
google-cloud-translate==3.20.1
flask-socketio==5.5.1
flask-cors==5.0.1
eventlet
gunicorn
